+++ src/evapotrans.c
@@ -22,22 +22,22 @@
 /*
  * Hamon method to estimate potential evapotranspiration
  */
-float hamon_et(float Tave, float latr, int jday){
+double hamon_et(double Tave, double latr, int jday){
 
     // Solar declination angle (radians)
     double d = 1 + 0.033*cos(2*M_PI*jday/365);
 
     // Sunset hour angle
-    float w = acos(-tan((double)latr)*tan(d)); 
+    double w = acos(-tan((double)latr)*tan(d)); 
 
     // Number of daylight hours
-    float N = 24*w/M_PI;
+    double N = 24*w/M_PI;
 
     // Saturation vapor pressure es at the mean daily temperature 
-    float e_s = 0.6108*exp(17.27*(double)Tave/((double)Tave+237.3));
+    double e_s = 0.6108*exp(17.27*(double)Tave/((double)Tave+237.3));
 
     // Saturation vapor density
-    float Pt = 216.7*e_s/(Tave+273.16);
+    double Pt = 216.7*e_s/(Tave+273.16);
 
     return HAMON_C*N*Pt/12;
 }
@@ -54,13 +54,13 @@ float hamon_et(float Tave, float latr, int jday){
  *  Potential evapotranspiration for multiple input data
  */
 /* int ets(evapot *evp, meteoin *metin, float lat, int ntimes){ */
-int ets(char *etmethod, float *tave, struct tm *timestamp, float lat, int ntimes, float *et){
+int ets(char *etmethod, double *tave, struct tm *timestamp, double lat, int ntimes, double *et){
 
     if (strcmp(etmethod, "hamon") == 0) {
         printf("ET computed using Hamon method\n");
         
         // Latitude to radians
-        float latr = lat * M_PI / 180.0;
+        double latr = lat * M_PI / 180.0;
         
         // Estimation of ET 
         /* float Tave; */
